Resolve an issue with authOptions
