+++ src/pages/api/restricted.js
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "./auth/[...nextauth]";
+import { authOptions } from "./auth/authOptions";
 
 export default async (req, res) => {
   const session = await getServerSession(req, res, authOptions);
